split big tests into smaller parts
